+++ ricaun.Revit.DB.Shape/Revit/Commands/ShapeCommand.cs
@@ -25,6 +25,8 @@ namespace ricaun.Revit.DB.Shape.Revit.Commands
                 var materialBlue = MaterialUtils.CreateMaterialBlue(document);
                 var materialWhite = MaterialUtils.CreateMaterialWhite(document);
                 var materialMagenta = MaterialUtils.CreateMaterial(document, Colors.Magenta);
+                var materialYellow = MaterialUtils.CreateMaterial(document, Colors.Yellow);
+                var materialCyan = MaterialUtils.CreateMaterial(document, Colors.Cyan);
 
                 var scale = 0.5;
 
@@ -37,11 +39,17 @@ namespace ricaun.Revit.DB.Shape.Revit.Commands
                 var cylinder = document.CreateDirectShape(ShapeCreator.CreateCylinder(XYZ.Zero, scale, materialId: materialBlue.Id));
                 cylinder.Location.Move(-(6 * scale) * XYZ.BasisX);
 
-                var cone = document.CreateDirectShape(ShapeCreator.CreateCone(XYZ.Zero, scale, materialId: materialWhite.Id));
+                var cone = document.CreateDirectShape(ShapeCreator.CreateCone(XYZ.Zero, scale, materialId: materialMagenta.Id));
                 cone.Location.Move(-(8 * scale) * XYZ.BasisX);
 
+                var prism = document.CreateDirectShape(ShapeCreator.CreatePrism(XYZ.Zero, scale, materialId: materialYellow.Id));
+                prism.Location.Move(-(10 * scale) * XYZ.BasisX);
+
+                var pyramid = document.CreateDirectShape(ShapeCreator.CreatePyramid(XYZ.Zero, scale, materialId: materialCyan.Id));
+                pyramid.Location.Move(-(12 * scale) * XYZ.BasisX);
+
                 var boxLines = document.CreateDirectShape(ShapeCreator.CreateBoxLines(XYZ.Zero, scale));
-                boxLines.Location.Move(-(10 * scale) * XYZ.BasisX);
+                boxLines.Location.Move(-(14 * scale) * XYZ.BasisX);
 
                 var arrowAxis = new XYZ(-1, -1, -1);
                 var arrow = document.CreateDirectShape(ShapeCreator.CreateArrow(arrowAxis, materialMagenta.Id));